requirements.txt
# BagHolderAI - Dependencies
# Exchange & Trading
ccxt==4.4.54
pandas==2.2.3
pandas-ta==0.3.14b0

# API Server
fastapi==0.115.6
uvicorn==0.34.0

# Database
supabase==2.11.0

# Telegram Bot
python-telegram-bot==21.9

# AI Sentinel
anthropic==0.43.0

# Utils
python-dotenv==1.0.1
httpx==0.28.1
schedule==1.2.2

# Dev & Testing
pytest==8.3.4
pytest-asyncio==0.25.0
